requirements.txt
numpy
torch
torchvision
Pillow
flask
matplotlib